+++ QKSMS/src/main/java/com/moez/QKSMS/common/DialogHelper.java
@@ -6,7 +6,7 @@ import com.google.gson.Gson;
 import com.moez.QKSMS.BuildConfig;
 import com.moez.QKSMS.R;
 import com.moez.QKSMS.data.Conversation;
-import com.moez.QKSMS.model.ChangeModel;
+import com.moez.QKSMS.mmssms.model.ChangeModel;
 import com.moez.QKSMS.ui.MainActivity;
 import com.moez.QKSMS.ui.base.QKActivity;
 import com.moez.QKSMS.ui.dialog.DefaultSmsHelper;
