+++ QKSMS/src/main/java/com/moez/QKSMS/common/NotificationManager.java
@@ -30,8 +30,8 @@ import com.moez.QKSMS.common.utils.ImageUtils;
 import com.moez.QKSMS.data.Contact;
 import com.moez.QKSMS.data.ContactHelper;
 import com.moez.QKSMS.data.Message;
-import com.moez.QKSMS.model.ImageModel;
-import com.moez.QKSMS.model.SlideshowModel;
+import com.moez.QKSMS.mmssms.model.ImageModel;
+import com.moez.QKSMS.mmssms.model.SlideshowModel;
 import com.moez.QKSMS.receiver.RemoteMessagingReceiver;
 import com.moez.QKSMS.ui.MainActivity;
 import com.moez.QKSMS.ui.messagelist.MessageItem;
