+++ QKSMS/src/main/java/com/moez/QKSMS/common/SmilHelper.java
@@ -29,15 +29,15 @@ import com.google.android.mms.MmsException;
 import com.google.android.mms.pdu_alt.PduBody;
 import com.google.android.mms.pdu_alt.PduPart;
 import com.moez.QKSMS.QKSMSApp;
-import com.moez.QKSMS.model.AudioModel;
-import com.moez.QKSMS.model.ImageModel;
-import com.moez.QKSMS.model.LayoutModel;
-import com.moez.QKSMS.model.MediaModel;
-import com.moez.QKSMS.model.RegionModel;
-import com.moez.QKSMS.model.SlideModel;
-import com.moez.QKSMS.model.SlideshowModel;
-import com.moez.QKSMS.model.TextModel;
-import com.moez.QKSMS.model.VideoModel;
+import com.moez.QKSMS.mmssms.model.AudioModel;
+import com.moez.QKSMS.mmssms.model.ImageModel;
+import com.moez.QKSMS.mmssms.model.LayoutModel;
+import com.moez.QKSMS.mmssms.model.MediaModel;
+import com.moez.QKSMS.mmssms.model.RegionModel;
+import com.moez.QKSMS.mmssms.model.SlideModel;
+import com.moez.QKSMS.mmssms.model.SlideshowModel;
+import com.moez.QKSMS.mmssms.model.TextModel;
+import com.moez.QKSMS.mmssms.model.VideoModel;
 import org.w3c.dom.events.EventTarget;
 import org.w3c.dom.smil.SMILDocument;
 import org.w3c.dom.smil.SMILElement;
