+++ QKSMS/src/main/java/com/moez/QKSMS/common/utils/MessageUtils.java
@@ -70,9 +70,9 @@ import com.moez.QKSMS.common.TempFileProvider;
 import com.moez.QKSMS.common.google.ThumbnailManager;
 import com.moez.QKSMS.common.google.UriImage;
 import com.moez.QKSMS.data.Contact;
-import com.moez.QKSMS.model.MediaModel;
-import com.moez.QKSMS.model.SlideModel;
-import com.moez.QKSMS.model.SlideshowModel;
+import com.moez.QKSMS.mmssms.model.MediaModel;
+import com.moez.QKSMS.mmssms.model.SlideModel;
+import com.moez.QKSMS.mmssms.model.SlideshowModel;
 import com.moez.QKSMS.ui.dialog.AsyncDialog;
 import com.moez.QKSMS.ui.messagelist.MessageColumns;
 import com.moez.QKSMS.ui.messagelist.MessageItem;
