+++ QKSMS/src/main/java/com/moez/QKSMS/mmssms/model/AudioModel.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.moez.QKSMS.model;
+package com.moez.QKSMS.mmssms.model;
 
 import android.content.ContentResolver;
 import android.content.Context;
