+++ QKSMS/src/main/java/com/moez/QKSMS/mmssms/model/ContentRestriction.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.moez.QKSMS.model;
+package com.moez.QKSMS.mmssms.model;
 
 import android.content.ContentResolver;
