+++ QKSMS/src/main/java/com/moez/QKSMS/mmssms/model/IModelChangedObserver.java
@@ -14,18 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.moez.QKSMS.model;
 
-public class ContentRestrictionFactory {
-    private static ContentRestriction sContentRestriction;
+package com.moez.QKSMS.mmssms.model;
 
-    private ContentRestrictionFactory() {
-    }
-
-    public static ContentRestriction getContentRestriction() {
-        if (null == sContentRestriction) {
-            sContentRestriction = new CarrierContentRestriction();
-        }
-        return sContentRestriction;
-    }
+public interface IModelChangedObserver {
+    void onModelChanged(Model model, boolean dataChanged);
 }
