+++ QKSMS/src/main/java/com/moez/QKSMS/ui/mms/MmsThumbnailPresenter.java
@@ -21,12 +21,12 @@ import android.content.Context;
 
 import com.moez.QKSMS.interfaces.SlideViewInterface;
 import com.moez.QKSMS.interfaces.ViewInterface;
-import com.moez.QKSMS.model.AudioModel;
-import com.moez.QKSMS.model.ImageModel;
-import com.moez.QKSMS.model.Model;
-import com.moez.QKSMS.model.SlideModel;
-import com.moez.QKSMS.model.SlideshowModel;
-import com.moez.QKSMS.model.VideoModel;
+import com.moez.QKSMS.mmssms.model.AudioModel;
+import com.moez.QKSMS.mmssms.model.ImageModel;
+import com.moez.QKSMS.mmssms.model.Model;
+import com.moez.QKSMS.mmssms.model.SlideModel;
+import com.moez.QKSMS.mmssms.model.SlideshowModel;
+import com.moez.QKSMS.mmssms.model.VideoModel;
 import com.moez.QKSMS.common.google.ItemLoadedCallback;
 import com.moez.QKSMS.common.google.ItemLoadedFuture;
 import com.moez.QKSMS.common.google.ThumbnailManager;
