+++ QKSMS/src/main/java/com/moez/QKSMS/ui/mms/Presenter.java
@@ -20,8 +20,8 @@ package com.moez.QKSMS.ui.mms;
 import android.content.Context;
 
 import com.moez.QKSMS.interfaces.ViewInterface;
-import com.moez.QKSMS.model.IModelChangedObserver;
-import com.moez.QKSMS.model.Model;
+import com.moez.QKSMS.mmssms.model.IModelChangedObserver;
+import com.moez.QKSMS.mmssms.model.Model;
 import com.moez.QKSMS.common.google.ItemLoadedCallback;
 
 /**
