+++ QKSMS/src/main/java/com/moez/QKSMS/ui/mms/SlideshowActivity.java
@@ -35,9 +35,9 @@ import com.android.mms.dom.smil.parser.SmilXmlSerializer;
 import com.google.android.mms.MmsException;
 import com.moez.QKSMS.R;
 import com.moez.QKSMS.common.SmilHelper;
-import com.moez.QKSMS.model.LayoutModel;
-import com.moez.QKSMS.model.RegionModel;
-import com.moez.QKSMS.model.SlideshowModel;
+import com.moez.QKSMS.mmssms.model.LayoutModel;
+import com.moez.QKSMS.mmssms.model.RegionModel;
+import com.moez.QKSMS.mmssms.model.SlideshowModel;
 import com.moez.QKSMS.ui.base.QKActivity;
 import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
